Major updates:
- Fixed wall boundary conditions: added the rotation back to global frame after performing operations in wall-frame. Previously, the state was left in wall-frame, leading to incorrect values being used for the boundary conditions

- caching: improved caching for state functions

- examples: added more examples

- meshes: better mesh names

+++ pyHype/fvm/Gradients/least_squares.py
@@ -160,8 +160,8 @@ def least_squares_9_point(Q, QE, QW, QN, QS,
                           stencilSW, stencilNW, stencilSE, stencilNE):
 
     # Initialize gradient arrays
-    dQdx = np.zeros((ny, nx, 4))
-    dQdy = np.zeros((ny, nx, 4))
+    dQdx = np.zeros((ny, nx, 4), dtype=float)
+    dQdy = np.zeros((ny, nx, 4), dtype=float)
 
     # ---------------------------------------------------------------------------------------
     # Corner cells
